refactor: extract TTS logic to TTSEngine class
